requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
decorator==5.1.1
Deprecated==1.2.13
Flask==2.2.2
frozenlist==1.3.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpath-ng==1.5.3
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
ply==3.11
pyparsing==3.0.9
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.20.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
